+++ R/ui.R
@@ -15,16 +15,18 @@ default_UI <- function(){
         headerPanel(.MCE$shinyGUI$title),
         
         if(is.finite(.MCE$design@max_time)){
-            h6(paste0('Time remaining: ', 
+            h6(paste0(.MCE$shinyGUI$time_remaining, 
                       formatTime(.MCE$design@max_time - sum(.MCE$person$item_time))))
         } else NULL,
         
         sidebarPanel(
             if(.MCE$shinyGUI$author != '') h4("Authors:") else NULL,
             if(.MCE$shinyGUI$author != '') h5(.MCE$shinyGUI$author) else NULL,
-            if(.MCE$shinyGUI$author != '') hr() else NULL,
-            helpText("\n", .MCE$shinyGUI$instructions[1L]),            
-            actionButton("Next", .MCE$shinyGUI$instructions[2L])
+            div(
+                if(.MCE$shinyGUI$author != '') hr() else NULL,
+                helpText("\n", .MCE$shinyGUI$instructions[1L]),            
+                actionButton("Next", .MCE$shinyGUI$instructions[2L])
+            )
         ),
         
         mainPanel(