+++ tests/GUI-tests/GUI-tests.R
@@ -10,11 +10,16 @@ questions <- c("Building CATs with mirtCAT is difficult.",
                "mirtCAT requires a substantial amount of coding.",
                "I would use mirtCAT in my research.")
 df <- data.frame(Question = questions, Option = options, Type = "radio")
-
-# forced
 results <- mirtCAT(df = df)
 summary(results)
-results <- mirtCAT(df = df, shinyGUI = list(stopApp = FALSE))
+
+# manual HTML tags, not forced
+df2 <- df
+questions <- c("Building <strong>CATs</strong> with mirtCAT is <br> <br> difficult.",
+               "mirtCAT requires a substantial amount of coding.",
+               "I would use mirtCAT in my research.")
+results <- mirtCAT(df = df2, shinyGUI=list(forced_choice = FALSE))
+summary(results)
 
 # correct answer scoring
 df2 <- df
@@ -23,6 +28,14 @@ df2$Mastery <- c(TRUE, FALSE, FALSE)
 results <- mirtCAT(df = df2)
 summary(results)
 
+# character coercion of R HTML constructors
+questions <- c(as.character(div(strong('Something'), br(), strong('Something'))),
+               "mirtCAT requires a substantial amount of coding.",
+               "I would use mirtCAT in my research.")
+df3 <- data.frame(Question = questions, Option = options, Type = "radio")
+
+results <- mirtCAT(df = df3)
+summary(results)
 
 #theme
 results <- mirtCAT(df = df, shinyGUI = list(theme = 'journal'))
@@ -60,10 +73,10 @@ choiceNames <- list(
         icon("calendar"),
         HTML("<p style='color:red;'>Red Text</p>"),
         "Normal text"),
-    NA, NA
+    NULL, NULL
 )
 choiceValues = list(
-    list("icon", "html", "text"), NA, NA)
+    list("icon", "html", "text"), NULL, NULL)
 results2 <- mirtCAT(df = df, shinyGUI = list(forced_choice = TRUE, 
                                              choiceNames=choiceNames,
                                              choiceValues=choiceValues))
@@ -152,28 +165,28 @@ df <- data.frame(Question = c("", "", "Just a standard stem."), Option = options
 results <- mirtCAT(df = df, shinyGUI = list(forced_choice = FALSE))
 
 # expressions
-df <- data.frame(Question = c('', 'tags$h1("My header")', 'tags$b("This text is bold.")'), 
+df <- data.frame(Question = c("", as.character(h1("My header")), 
+                              as.character(tags$b("This text is bold."))), 
                  Option = options, Type = "radio",
-                 Stem = c('Math-stem.html', '', ''),
-                 StemExpression = c(FALSE, TRUE, TRUE))
+                 Stem = c('Math-stem.html', '', ''))
 results <- mirtCAT(df = df, shinyGUI = list(forced_choice = FALSE))
 
 # divs
-df <- data.frame(Question = c('div(HTML("Here is <strong>one</strong> way to insert <em>arbitrary</em> HTML."))', 
-                              'div(tags$style("#text { font-size: 35px; height: 200px; overflow: auto; }"), 
-                                         div(id = "text", paste(names(tags), collapse = ", ")))', 
-                              'div(tags$style("#text { font-size: 20px; height: 65px; overflow: auto; }"), 
-                                         div(id = "text", paste(names(tags), collapse = ", ")))'), 
-                 Option = options, Type = "radio",
-                 Stem = c('', '', ''),
-                 StemExpression = rep(TRUE, 3))
+df <- data.frame(Question = sapply(list(div(HTML("Here is <strong>one</strong> way to insert <em>arbitrary</em> HTML.")), 
+                                        div(tags$style("#text { font-size: 35px; height: 200px; overflow: auto; }"), 
+                                            div(id = "text", paste(names(tags), collapse = ", "))), 
+                                        div(tags$style("#text { font-size: 20px; height: 65px; overflow: auto; }"), 
+                                            div(id = "text", paste(names(tags), collapse = ", ")))), as.character)
+                     , 
+                 Option = options, Type = "radio")
 results <- mirtCAT(df = df, shinyGUI = list(forced_choice = FALSE))
 
 # table panels
-df <- data.frame(Question = c('tabsetPanel(tabPanel("Panel 1", "some text"), tabPanel("Panel 2", "some more text"))', 
-                              'tags$h1("My header")', 'tags$b("This text is bold.")'), 
-                 Option = options, Type = "radio",
-                 StemExpression = c(TRUE, TRUE, TRUE))
+df <- data.frame(Question = c(as.character(tabsetPanel(tabPanel("Panel 1", "some text"), 
+                                                       tabPanel("Panel 2", "some more text"))),
+                              'Something', 
+                              as.character(tags$b("This text is bold."))),
+                 Option = options, Type = "radio")
 results <- mirtCAT(df = df, shinyGUI = list(forced_choice = FALSE))
 
 # custom choices
@@ -198,14 +211,13 @@ results <- mirtCAT(df = df, shinyGUI = list(forced_choice = FALSE),
 # audio/video
 dirname <- paste0(getwd(), '/www')
 shiny::addResourcePath('www', dirname)
-df <- data.frame(Question = c('', 
-                              'tags$audio(src = "www/clip.mp3", type = "audio/mp3",
-                                    autoplay = TRUE, controls = TRUE)', 
-                              'tags$video(src = "www/vid.mp4", type = "video/mp4",
-                                    controls = TRUE, height=260, width=260)'), 
-                 Option = options, Type = "radio",
-                 Stem = c('Math-stem.html', '', ''),
-                 StemExpression = c(FALSE, TRUE, TRUE))
+df <- data.frame(Question = c("", 
+                              as.character(tags$audio(src = "www/clip.mp3", type = "audio/mp3",
+                                                      autoplay = TRUE, controls = TRUE)),
+                              as.character(tags$video(src = "www/vid.mp4", type = "video/mp4",
+                                         controls = TRUE, height=260, width=260))),
+                              Option = options, Type = "radio",
+                 Stem = c('Math-stem.html', '', ''))
 results <- mirtCAT(df = df, shinyGUI = list(forced_choice = FALSE))
 
 # checkbox input